+++ src/OptFlowCpp/cpu_backend/linalg/ArithmeticChained.h
@@ -32,7 +32,7 @@ namespace cpu_backend
 			auto in_c = std::dynamic_pointer_cast<Array<InnerTyp, DimCount>>(c);
 
 			#pragma omp parallel for
-			for (size_t i = 0; i < size; i++)
+			for (int i = 0; i < size; i++)
 			{
 				(*out)[i] = (*in_a)[i] * (*in_b)[i] + (*in_c)[i];
 			}
@@ -52,7 +52,7 @@ namespace cpu_backend
 			auto in_d = std::dynamic_pointer_cast<Array<InnerTyp, DimCount>>(d);
 
 			#pragma omp parallel for
-			for (size_t i = 0; i < size; i++)
+			for (int i = 0; i < size; i++)
 			{
 				(*out)[i] = (*in_a)[i] * (*in_b)[i] + (*in_c)[i] * (*in_d)[i];
 			}
@@ -70,7 +70,7 @@ namespace cpu_backend
 			auto in_b = std::dynamic_pointer_cast<Array<InnerTyp, DimCount>>(b);
 
 			#pragma omp parallel for
-			for (size_t i = 0; i < size; i++)
+			for (int i = 0; i < size; i++)
 			{
 				(*out)[i] = alpha * (*in_a)[i] + (*in_b)[i];
 			}
